Update audited byte-accuracy source recovery
Callee decls corrected (signature only, name unchanged):
- action_guard_setup_find_position @ 0x15900 (<common>)
    int action_guard_setup_find_position(int actor_handle, short param_2, char *state_data)
 -> int action_guard_setup_find_position(int actor_handle, int param_2, char *state_data)
- rasterizer_windows_begin @ 0x17c8c0 (<common>)
    int rasterizer_windows_begin(void)
 -> void rasterizer_windows_begin(void)
- rasterizer_window_begin @ 0x17c8d0 (<common>)
    int rasterizer_window_begin(window_parameters_t *a1)
 -> void rasterizer_window_begin(window_parameters_t *a1)
- __ascii_stricmp @ 0x1e3a50 (<common>)
    void __ascii_stricmp(void)
 -> int __ascii_stricmp(const char *a, const char *b)
- FUN_000dc8c0 @ 0xdc8c0 (<common>)
    int16_t FUN_000dc8c0(int16_t state@<eax>)
 -> int FUN_000dc8c0(int16_t state@<eax>)

Coverage: 0.0% (0/0 symbols)

+++ src/halo/interface/first_person_weapons.c
@@ -50,7 +50,7 @@ void FUN_000dc7f0(void)
 /* Map a first-person weapon state to an animation graph index (0xdc8c0).
  * Pure lookup table: 24 states (0..23) map to animation indices; any
  * out-of-range state returns -1. */
-int16_t FUN_000dc8c0(int16_t state)
+int FUN_000dc8c0(int16_t state)
 {
   switch (state) {
   case 0:
